Clickable/ bodyparts put in AngularModels
